Got all tests to pass.
Some were failing because we now run the analyzer using the actual compiler driver so more diagnostics are reported than if we cherry pick nodes to pass in.  These were originally marked as false positives, but I changed the tests to pass and have this be the new behavior.

Also, the Foreach analyzer reports things that don't make sense to me.  An array containing a value type does not cause an additional allocation.  It won't even cause an enumerable to be allocated since the compiler re-writes this to be a for loop.  The same goes for List, it will copy the array, and thats and allocation, but it not avoidable or unexpected.  As far as I could tell, this analyzer was not reporting anthing useful so I removed it and its tests.