Analyze hot paths
Added the PerformanceCriticalAttribute to its own assembly

Members marked with this attribute will enable analysis for all rules within that method.

If one member is marked, no unmarked members of a class will be analyzed.

If a class does not have any method with the attribute, analysis is done on every method (old behavior).